pyproject.toml
[tool.poetry]
name = "visible-ideas-homework"
version = "0.1.0"
description = ""
authors = ["binarycat <artur.rakhmatulin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
